changes for tpm simulator instructions ans scripts:

+++ vm_model_tools/examples/scenario1/clean-tpm-simulator.sh
@@ -36,18 +36,33 @@ else
 fi
 TPM_SUPPORT_DIR=$CERTIFIER_ROOT/src/tpm2
 
+if [[ ! -v XDG_CONFIG_HOME ]]; then
+  echo "Using export XDG_CONFIG_HOME=/home/jlm/.config"
+  export XDG_CONFIG_HOME="/home/jlm/.config"
+fi
+if [[ ! -e $XDG_CONFIG_HOME ]]; then
+  echo "$XDG_CONFIG_HOME does not exist"
+  exit
+fi
+
 echo " "
 echo "Certifier root: $CERTIFIER_ROOT"
 echo "TPM support directory: $TPM_SUPPORT_DIR"
 
-  echo " "
-  echo "cleanup-tpm-simulator"
+echo " "
+echo "cleanup-tpm-simulator"
 
-  # kill the server
-  cleanup-stale-procs
+# kill the server
+cleanup-stale-procs
 
-  # remove the files
-  rm seal_hierarchy.bin quote_hierarchy.bin
+# remove the files
+rm seal_hierarchy.bin quote_hierarchy.bin
+
+# clear tpm state
+pushd $XDG_CONFIG_HOME/mytpm1
+  rm ./* || true
+  rm ./.lock || true
+popd
 
 echo "Done"
 exit
